resources: extracted unit tests for Find, Types, and Names to extras/resources
